requirements.txt
openai==0.28
groq
google-generativeai
python-dotenv
instructor
colorama